requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.22
tqdm
matplotlib
pandas

speechbrain==0.5.15
datasets
torch_audiomentations
wandb
torchaudio==2.1.0
pillow
gdown
Unidecode
inflect
nltk
hparams
TextGrid
pyworld